[Task 9] file restructure
- project/cfpq_utils/hellings_algo.py
- project/cfpq_utils/cfpq_algo.py
- tests/test_cfpq_hellings.py
